Add a convenience initializer to create a new managed object, inserting it into the specified managed object context.
This method calls the designated initializer
`init(entity:insertInto:)` using the class name to create
the entity description. This avoids a problem, often seen
in unit tests, where the convenience initializer
`init(context:)` generates a warning that multiple
`NSEntityDescriptions` claim the `NSManagedObject` subclass
when the managed object model is loaded more than once.

Note: This method assumes the class name matches the
entity name in the model. It will crash if the entity
cannot be found in the managed object model.